source/requirements.txt
gunicorn==20.1.0
django==4.1
python-decouple==3.6
psycopg2-binary==2.9.3
django-multiselectfield==0.1.12
pillow==9.2.0
django-phonenumber-field[phonenumberslite]
factory-boy==3.2.1
coverage==6.4.4
django-jazzmin==2.5.0
whitenoise==6.2.0